requirements.txt
flask
psycopg2-binary
python-telegram-bot
